tradingeconomics/requirements.txt
pandas
python_dateutil==2.8.2
requests==2.28.2
websocket_client==1.3.3
